+++ src/dune_rules/coq_sources.ml
@@ -12,12 +12,19 @@ type t =
         (* [directories] is used to compute the include paths for Coq's native
            mode *)
   ; extract : Coq_module.t Loc.Map.t
+  ; rev_map : [ `Theory of Theory.t
+              | `Extraction of Extraction.t ] Coq_module.Map.t
   }
+
+let find_module ~source t =
+  let f m = Path.Build.equal source (Coq_module.source m) in
+  Coq_lib_name.Map.Multi.find_elt t.libraries ~f
 
 let empty =
   { libraries = Coq_lib_name.Map.empty
   ; directories = Coq_lib_name.Map.empty
   ; extract = Loc.Map.empty
+  ; rev_map = Coq_module.Map.empty
   }
 
 let coq_modules_of_files ~dirs =
@@ -54,7 +61,7 @@ let of_dir (d : _ Dir_with_dune.t) ~include_subdirs ~dirs =
   check_no_unqualified include_subdirs;
   let modules = coq_modules_of_files ~dirs in
   List.fold_left d.data ~init:empty ~f:(fun acc -> function
-    | Coq_stanza.Theory.T coq ->
+    | Theory.T coq ->
       let modules =
         Coq_module.eval ~dir:d.ctx_dir coq.modules ~standard:modules
       in
@@ -65,9 +72,13 @@ let of_dir (d : _ Dir_with_dune.t) ~include_subdirs ~dirs =
       let libraries =
         Coq_lib_name.Map.add_exn acc.libraries (snd coq.name) modules
       in
-      { acc with directories; libraries }
-    | Coq_stanza.Extraction.T extract ->
-      let loc, prelude = extract.prelude in
+      let rev_map =
+        List.fold_left modules ~init:acc.rev_map ~f:(fun acc m ->
+          Coq_module.Map.add_exn acc m (`Theory coq))
+      in
+      { acc with directories; libraries; rev_map }
+    | Extraction.T extr ->
+      let loc, prelude = extr.prelude in
       let m =
         match
           List.find modules ~f:(fun m ->
@@ -78,6 +89,10 @@ let of_dir (d : _ Dir_with_dune.t) ~include_subdirs ~dirs =
           User_error.raise ~loc
             [ Pp.text "no coq source corresponding to prelude field" ]
       in
-      let extract = Loc.Map.add_exn acc.extract extract.buildable.loc m in
-      { acc with extract }
+      let extract = Loc.Map.add_exn acc.extract extr.buildable.loc m in
+      let rev_map = Coq_module.Map.add_exn acc.rev_map m (`Extraction extr) in
+      { acc with extract; rev_map }
     | _ -> acc)
+
+let lookup_module t m =
+  Coq_module.Map.find t.rev_map m